Handle ignored and abandoned statuses (#10)
These statuses are not handled at the moment.
When a job completes with either of them the runner is not deprovisioned.
This results in the runner and the VM taking resources from the Orka Cluster.
Handle ignored and abandoned statuses as cancelled. This ensures the runner and the VM are deprovisioned.
And the cluster has the resources to run other jobs with other runners.