Reverse the order of creating the runner and VM
Creating the runner after the VM ensures that the runner does not stay in the GitHub UI in an "unknown/not ready" state.
While this does not cause any issues, it might be misleading for users as they might think the runner is orhpaned.
The new logic creates the runner only after there is a running VM and it has a valid IP address.