Handle missing jobId
If a jobID is missing (GitHub change their API again), use a default value.
This default value prevents runner provisioning from being cancelled.

What happens when a jobId is missing and the job is cancelled?

If the runner is already provisioned, it will be deprovisioned.
If the runner is not provisioned, it will not be provisioned. It will stay in idle state and will pick the next available job.
This is ok as cancelling jobs before a runner is provisioned does not happen often and even if it happens, the runner is going to be available for the next job.