Remove the deprovisioning code
If a job has a runner assigned, it will be deprovisioned automatically.
If a job has no runner assigned, the job does not have a runner name.
This means we cannot deprovision the runner as it is still being created.

This can happen only if a job is cancelled while the runner is being created.
The result would be a runner in idle state and will pick the next available job.